Revert "Merge pull request #71 from freeformsystems/master"
This reverts commit e7df1f0d254836f6ac7760a2727b4bd342d52ecb, reversing
changes made to c3fd67e94094e3e366e958d746c68c2903756294.

Commit was introduce as a aesthetic change when using `-Kutf8 -Tutf8`.

Quoting from the issue that led to the PR:

> This may seem a little strange, it's more aesthetic then anything
> else. When troff is run with -Kutf8 -Tutf8 to ensure that utf8
> characters appear correctly then the bullets generated by \(bu appear as
> small squares (I'm using 13pt Pragmata Pro).

However, in `bundler` we're not using `-Tutf8`, but `-Tascii`, and this
results in weird characters for list items as reported [here](https://github.com/cli-kit/cli-command/issues/78).

Since this change never made it into any `ronn` release, I think
reverting it could encourage adoption of `ronn-ng` as a drop-in
replacement of `ronn`. At least for us :)