[doc] Fix some bad markup in `ronn.1.ronn` that was causing bad list rendering
Looks like the missing trailing colon (":") on one item in a definition list was causing the whole list to be mis-detected as a bullet list instead. This change fixes the rendering of the ronn.1 man page itself. But I think there should be an additional fix that has ronn issue a warning or error when it encounters a list with inconsistent item types and there's some ambiguity as to what type of list it is.

See: https://github.com/apjanke/ronn-ng/issues/94