[code] add rbenv support, notes on dev envs, code formatting
Adds .ruby-version to the repo, because as an application, Ronn targets a specific Ruby version (like 2.7.x), and the gem versions in Gemfile.lock also depend on that. So I want it included in the repo to reflect that, and make things more reproducible and convenient across different repo clones.