[deps] require Ruby 2.7+ instead of 2.4+, bc nokogiri needs it
We need nokogiri >= 1.14.3 to get the right behavior for namespace-looking tag names with ":" in them, and nokogiri 1.14.3 requires ruby >= 2.7. Ruby 2.7+ is all I want to work and test on for other reasons, anyway. Users who really want to deploy on older Rubies can edit the gemspec file.

But if I'm being honest, really my main reason for doing this is to make Rubocopy happy.