tweak `ronn --version` output: remove github URL line, make it "ronn <version> (Ronn-NG)"
Removed the "http://github.com/..." line for the release website because I just don't like it. And also it's wrong if you're running from a repo working copy or branch with changes.

Changed the main line from "Ronn-NG <version>" to "ronn <version> (Ronn-NG)" because I think it's conventional to put the command name first. Ronn-NG is a whole project which might in theory produce multiple commands with independent versions.

Will maybe remove the "(Ronn-NG)" eventually if this project establishes itself as the canonical current Ronn.