Sniff H2 sections when ingesting documents
Documents may already have NAME sections which already have name and
tagline, so we should try that before looking at the H1 header.

The sections may use the hyphenated form, which we can split up as
name and tagline, or they may just have the name of the page.  We can
handle both.